+++ polybench-code/linear-algebra/solvers/cholesky/cholesky.c
@@ -17,7 +17,7 @@ static
 void init_array(int n,
 		DATA_TYPE POLYBENCH_2D(A,N,N,n,n))
 {
-  int i, j;
+  long long i, j;
 
   for (i = 0; i < n; i++)
     {
@@ -54,7 +54,7 @@ void print_array(int n,
 		 DATA_TYPE POLYBENCH_2D(A,N,N,n,n))
 
 {
-  int i, j;
+  long long i, j;
 
   POLYBENCH_DUMP_START;
   POLYBENCH_DUMP_BEGIN("A");
@@ -74,7 +74,7 @@ static
 void kernel_cholesky(int n,
 		     DATA_TYPE POLYBENCH_2D(A,N,N,n,n))
 {
-  int i, j, k;
+  long long i, j, k;
 
 
 #pragma scop
